requirements.txt
ansible==2.7.9
ansible-lint==4.1.0
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.1
bcrypt==3.1.6
beautifulsoup4==4.7.1
bleach==3.1.0
Brlapi==0.7.0
btrfsutil==1.1.0
CacheControl==0.12.5
cffi==1.12.2
chardet==3.0.4
Click==7.0
cmarkgfm==0.4.2
colorama==0.4.1
cryptography==2.6.1
decorator==4.3.2
distlib==0.2.8
distro==1.4.0
Django==2.2
django-netfields==0.10.0
djangorestframework==3.9.2
docutils==0.14
Flask==1.0.2
gpg==1.13.0
hkp4py==0.2.3.0
html5lib==1.0.1
idna==2.8
iotop==0.6
isc==2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
ldap3==2.6
lensfun==0.3.2
libvirt-python==5.0.0
lockfile==0.12.2
louis==3.8.0
lxml==4.3.3
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
meld==3.20.1
msgpack==0.6.1
multivault==0.7.0.0
netaddr==0.7.19
numpy==1.16.2
packaging==19.0
paramiko==2.4.2
pep517==0.5.0
pkginfo==1.5.0.1
ply==3.11
progress==1.5
psycopg2==2.7.7
pwquality==1.4.0
pyasn1==0.4.5
PyAudio==0.2.11
pycairo==1.18.0
pycodestyle==2.4.0
pycparser==2.19
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.32.0
pylint==2.2.2
PyNaCl==1.3.0
pyparsing==2.3.1
PyQt5==5.12.1
PyQt5-sip==4.19.15
PyQtWebEngine==5.12.1
pyserial==3.4
python-dateutil==2.8.0
pytoml==0.1.20
pytz==2018.9
pyxdg==0.26
PyYAML==3.13
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
retrying==1.3.3
ruamel.yaml==0.15.91
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
stevedore==1.30.0
team==1.0
tqdm==4.28.1
twine==1.13.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
voluptuous==0.11.5
webencodings==0.5.1
Werkzeug==0.15.1
wrapt==1.11.1
youtube-dl==2019.3.18
